requirements.txt
numpy
fastapi
uvicorn
langchain==1.0.3
langchain-core==1.0.3
langchain-google-genai==3.0.1
google-ai-generativelanguage==0.9.0
pydantic
python-dotenv